+++ decidim-initiatives/app/controllers/decidim/initiatives/admin/component_permissions_controller.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Decidim
+  module Initiatives
+    module Admin
+      # Controller that allows managing the Initiative's Component
+      # permissions in the admin panel.
+      class ComponentPermissionsController < Decidim::Admin::ComponentPermissionsController
+        include InitiativeAdmin
+      end
+    end
+  end
+end