+++ decidim-initiatives/db/migrate/20171017094911_add_scoped_type_to_initiative.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+class AddScopedTypeToInitiative < ActiveRecord::Migration[5.1]
+  def change
+    add_column :decidim_initiatives,
+               :scoped_type_id, :integer, index: true
+  end
+end