+++ decidim-initiatives/lib/decidim/initiatives/admin_engine.rb
@@ -1,0 +1,95 @@
+# frozen_string_literal: true
+
+require "rails"
+require "active_support/all"
+require "decidim/core"
+
+module Decidim
+  module Initiatives
+    # Decidim's Assemblies Rails Admin Engine.
+    class AdminEngine < ::Rails::Engine
+      isolate_namespace Decidim::Initiatives::Admin
+
+      paths["db/migrate"] = nil
+
+      routes do
+        resources :initiatives_types, except: :show do
+          resources :initiatives_type_scopes, except: [:index, :show]
+        end
+
+        resources :initiatives, only: [:index, :show, :edit, :update], param: :slug do
+          member do
+            get :send_to_technical_validation
+            post :publish
+            delete :unpublish
+            delete :discard
+            get :export_votes
+            post :accept
+            delete :reject
+          end
+
+          resources :attachments, controller: "initiative_attachments"
+
+          resources :committee_requests, only: [:index] do
+            member do
+              get :approve
+              delete :revoke
+            end
+          end
+        end
+
+        scope "/initiatives/:initiative_slug" do
+          resources :components do
+            resource :permissions, controller: "component_permissions"
+            member do
+              put :publish
+              put :unpublish
+            end
+            resources :exports, only: :create
+          end
+        end
+
+        scope "/initiatives/:initiative_slug/components/:component_id/manage" do
+          Decidim.component_manifests.each do |manifest|
+            next unless manifest.admin_engine
+
+            constraints CurrentComponent.new(manifest) do
+              mount manifest.admin_engine, at: "/", as: "decidim_admin_initiative_#{manifest.name}"
+            end
+          end
+        end
+      end
+
+      initializer "admin_decidim_initiatives.assets" do |app|
+        app.config.assets.precompile += %w(
+          admin_decidim_initiatives_manifest.js
+        )
+      end
+
+      initializer "decidim_assemblies.inject_abilities_to_user" do |_app|
+        Decidim.configure do |config|
+          config.admin_abilities += %w(
+            Decidim::Initiatives::Abilities::Admin::CommitteeUserAbility
+            Decidim::Initiatives::Abilities::Admin::InitiativeUserAbility
+            Decidim::Initiatives::Abilities::Admin::CommitteeAdminAbility
+            Decidim::Initiatives::Abilities::Admin::InitiativeAdminAbility
+            Decidim::Initiatives::Abilities::Admin::InitiativeTypeAbility
+            Decidim::Initiatives::Abilities::Admin::AttachmentsAbility
+            Decidim::Initiatives::Abilities::Admin::ComponentsAbility
+          )
+        end
+      end
+
+      initializer "decidim_assemblies.admin_menu" do
+        Decidim.menu :admin_menu do |menu|
+          menu.item I18n.t("menu.initiatives", scope: "decidim.admin"),
+                    decidim_admin_initiatives.initiatives_path,
+                    icon_name: "chat",
+                    position: 3.7,
+                    active: :inclusive,
+                    if: can?(:index, Decidim::Initiative)
+        end
+      end
+    end
+  end
+end